split Configuration to BasicConfiguration and Configuration
